+++ package.json
@@ -7,7 +7,7 @@
         "serve": "webpack serve --open --mode development",
         "dev": "webpack --mode development",
         "build": "cross-env NODE_ENV=production webpack --mode production",
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "test": "jest",
         "format": "prettier --write .",
         "lint": "stylelint **/**.scss",
         "prepare": "husky install"
@@ -27,8 +27,14 @@
     "author": "",
     "license": "ISC",
     "devDependencies": {
-        "@babel/core": "^7.21.8",
-        "@babel/preset-env": "^7.21.5",
+        "@babel/core": "^7.22.1",
+        "@babel/preset-env": "^7.22.4",
+        "@babel/preset-typescript": "^7.21.5",
+        "@jest/globals": "^29.5.0",
+        "@testing-library/dom": "^9.3.0",
+        "@types/jest": "^29.5.1",
+        "@types/node": "^20.2.5",
+        "babel-jest": "^29.5.0",
         "babel-loader": "^9.1.2",
         "copy-webpack-plugin": "^11.0.0",
         "cross-env": "^7.0.3",
@@ -46,6 +52,8 @@
         "imagemin-jpegtran": "^7.0.0",
         "imagemin-optipng": "^8.0.0",
         "imagemin-svgo": "^10.0.1",
+        "jest": "^29.5.0",
+        "jest-environment-jsdom-global": "^4.0.0",
         "lint-staged": "^13.2.2",
         "mini-css-extract-plugin": "^2.7.5",
         "prettier": "2.8.8",
@@ -57,8 +65,11 @@
         "stylelint-order": "^6.0.3",
         "stylelint-prettier": "^3.0.0",
         "svgo": "^3.0.2",
+        "ts-jest": "^29.1.0",
         "ts-loader": "^9.4.2",
         "ts-migrate": "^0.1.35",
+        "ts-node": "^10.9.1",
+        "tslib": "^2.5.2",
         "typescript": "^5.0.4",
         "webpack": "^5.82.0",
         "webpack-cli": "^5.0.2",
