+++ tsconfig.json
@@ -13,7 +13,7 @@
         // "disableReferencedProjectLoad": true,             /* Reduce the number of projects loaded automatically by TypeScript. */
 
         /* Language and Environment */
-        "target": "es6" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
+        "target": "esnext" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
         "lib": ["dom", "es5", "es2016", "es2015.collection", "es2015.iterable"],
         /* Specify a set of bundled library declaration files that describe the target runtime environment. */ // "jsx": "preserve",                                /* Specify what JSX code is generated. */
         // "experimentalDecorators": true,                   /* Enable experimental support for legacy experimental decorators. */
@@ -27,9 +27,9 @@
         // "moduleDetection": "auto",                        /* Control what method is used to detect module-format JS files. */
 
         /* Modules */
-        "module": "es6" /* Specify what module code is generated. */,
+        "module": "esnext" /* Specify what module code is generated. */,
         // "rootDir": "./",                                  /* Specify the root folder within your source files. */
-        // "moduleResolution": "node10",                     /* Specify how TypeScript looks up a file from a given module specifier. */
+        "moduleResolution": "node" /* Specify how TypeScript looks up a file from a given module specifier. */,
         "baseUrl": "./src" /* Specify the base directory to resolve non-relative module names. */,
         // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
